:building_construction: rename module to geometry
